Add recursive scraping of "Two Peas" main dishes
The spider will look for the next page button (">") at the bottom of the
recipe, and add it to the pages to scrape. This allows the spider to
cover the entire main dishes category by starting from the first page.

Seems to introduce the unintended effect of scraping non-"main dish"
pages from earlier recipes, which may be a bug.

Resolves: #1